+++ CHANGELOG.md
@@ -7,7 +7,8 @@
 - SDK: added IotTimeSeriesAggregateClient to the SDK
 - SDK: added SpectrumAnalysisClient to the SDK
 - CLI Command: mc spectrum-analysis: performs spectrum analysis on a sound
-- CLI Command: mc list-files: lists all iot files stored with the asset
+- CLI Command: mc list-files: lists iot files stored with the asset
+- CLI Command: mc list-assets: lists assets in the tenant
 
 ## Bugfix 3.6.0